If there is a successful network response, then cache those contacts. Those are used in the Contacts UI, and they are needed again to check if any local contacts are using the app. I check that against the remote db and display in in the search screen accordingly. So I added list to hold contacts cache for the search screen. The flow continues on and displays what is needed in the UI also.
Started work on search UI to begin displaying that data. Added fetchUsersByPhoneNumber to query the remote db for users. And added helper functions to CombinedContactsRepository as needed. There is a companion object there also, where the cache of contacts is held.

Removed injected SearchResult as it was not being used along with some interface functions that I am not using at the moment. Left a todo in ContactsQueryHelper just to go back later to make sure no internet is being handled the way I want.